Refactoring __init__ for poes
